move merger da raiz para camada transform
